+++ pkg/dbt_cloud/credential.go
@@ -1,0 +1,30 @@
+package dbt_cloud
+
+import (
+	"fmt"
+	"net/http"
+)
+
+func (c *Client) DeleteCredential(credentialId, projectId string) (string, error) {
+	req, err := http.NewRequest(
+		"DELETE",
+		fmt.Sprintf(
+			"%s/v3/accounts/%d/projects/%s/credentials/%s/",
+			c.HostURL,
+			c.AccountID,
+			projectId,
+			credentialId,
+		),
+		nil,
+	)
+	if err != nil {
+		return "", err
+	}
+
+	_, err = c.doRequest(req)
+	if err != nil {
+		return "", err
+	}
+
+	return "", err
+}